RELEASE-VERSION:0.0.4
【2020.03.16】
后台
-a 提供feign解决方案，并集成ribbon提供负载均衡
-a 字典组件功能，并支持缓存
-a 增加RSA工具，实现基于注解的自动解密接口参数（支持param和body）
-a 增加基础邮件推送功能
-a 标准查询接口支持字典相关高级查询
-u 生成器对feign的支持
-u 生成器对vue-modal的支持
-u 生成器对vue-list(tree)的支持
-u 优化重复属性校验方法
-f 修复自定义异常入库丢失的问题
-f 修复刷新token前端无法获取的问题
-f 修复redis监控数据的异常

前台
-a log工具类
-a 新增角色验证属性v-role和方法hasRole
-a 基础表格基础按钮组，提供刷新、导入、导出（支持自定义列和自定义行）
-a list增加排序参数
-a redis监控完成 && 服务器信息监控完成
-a 新增图片选择控件 && 字典组件完成
-a 新增角色管理基础页面 && 字典和字典详情管理界面
-u modal基础方法抽离
-u 权限校验支持数组
-u 基础表格优化，支持自定义列
-u 搜索组件支持字典搜索
-u token自动刷新
-u 操作按钮批量操作列表也支持扩展
-u 基础表格slot插槽实现，表头和内容都允许以slot插槽的方式实现
-u 用户编辑重复校验完善
-f 菜单编辑树状选择框异常修复
-f 刷新出现的横向白框修复

其他

【计划】

1.完成权限管理功能（授权/角色）
2.系统运维监控功能（tomcat/jvm/request）
3.字典管理功能
4.邮件模板管理&邮件推送

+++ T4Cloud-System/src/main/java/com/t4cloud/t/system/controller/SysRolePermissionController.java
@@ -14,6 +14,7 @@ import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
@@ -42,7 +43,7 @@ public class SysRolePermissionController extends T4Controller<SysRolePermission,
     /**
      * 全部列表 角色权限表
      */
-    @AutoLog(value = "角色权限表-全部列表", operateType = 4)
+    @AutoLog(value = "角色权限表-全部列表")
     @GetMapping("/list")
     @RequiresPermissions("system:SysRolePermission:VIEW")
     @ApiOperation(position = 1, value = "角色权限表-全部列表", notes = "传入sysRolePermission")
@@ -57,29 +58,26 @@ public class SysRolePermissionController extends T4Controller<SysRolePermission,
      */
     @AutoLog(value = "保存 角色授权", operateType = 1)
     @PutMapping("/save")
-    @RequiresPermissions("system:SysRolePermission:ADD")
-    @ApiOperation(position = 4, value = "角色权限表-新增", notes = "传入sysRolePermission")
+    @RequiresPermissions("system:SysRolePermission:EDIT")
+    @ApiOperation(position = 4, value = "保存 角色授权", notes = "传入sysRolePermission")
     @PermissionCacheEvict
-    public R save(@Valid @RequestBody SysRolePermission sysRolePermission) {
+    public R<?> save(@Valid @RequestBody SysRolePermission sysRolePermission, BindingResult bindingResult) {
         //查询之前的所有权限
         List<SysRolePermission> sysRolePermissions = service.lambdaQuery().eq(SysRolePermission::getRoleId, sysRolePermission.getRoleId()).list();
         List<String> oldPermissionIds = sysRolePermissions.stream().map(SysRolePermission::getPermissionId).collect(Collectors.toList());
         //新的权限
         List<String> newPermissionIds = Arrays.asList(sysRolePermission.getPermissionIds().split(","));
-
         //处理新增的
         List<String> addPermissionIds = newPermissionIds.stream().filter(item -> StringUtils.isNotEmpty(item) && !oldPermissionIds.contains(item)).collect(Collectors.toList());
-        List<SysRolePermission> addPermissionIdList = new ArrayList<SysRolePermission>();
+        List<SysRolePermission> addPermissionIdList = new ArrayList<>();
         for (String permissionId : addPermissionIds) {
             SysRolePermission rolepms = new SysRolePermission().setRoleId(sysRolePermission.getRoleId()).setPermissionId(permissionId);
             addPermissionIdList.add(rolepms);
         }
         service.saveBatch(addPermissionIdList);
-
         //处理删除的
         List<String> deletePermissionIds = oldPermissionIds.stream().filter(item -> StringUtils.isNotEmpty(item) && !newPermissionIds.contains(item)).collect(Collectors.toList());
-        service.removeByIds(deletePermissionIds);
-
+        service.removeByIds(sysRolePermissions.stream().filter(item -> deletePermissionIds.contains(item.getPermissionId())).map(SysRolePermission::getId).collect(Collectors.toList()));
         return R.ok("角色授权表-保存成功");
     }
